pyproject.toml
[tool.poetry]
name = "codereviewai"
version = "0.1.0"
description = ""
authors = ["dzhusjr <romeodzhus3304@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.5"
uvicorn = "^0.32.1"
requests = "^2.32.3"
openai = "^1.55.0"
pydantic = "^2.10.1"
pytest = "^8.3.3"
httpx = "^0.27.2"
pytest-asyncio = "^0.24.0"
aiohttp = "^3.11.7"
python-dotenv = "^1.0.1"


[tool.poetry.group.dev.dependencies]
black = "^24.10.0"
flake8 = "^7.1.1"
isort = "^5.13.2"
mypy = "^1.13.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
